requirements.txt
langchain>=0.1.12
langchain-community>=0.0.28
python-dotenv==1.0.1
ollama>=0.5.1
notion-client==2.2.1
faiss-cpu==1.7.4
sentence-transformers>=4.1.0
scikit-learn==1.4.1.post1
scipy==1.12.0
transformers>=4.41.0
flask==3.0.2